+++ Runtime/wechat-default/unity-namespace.js
@@ -173,7 +173,7 @@ bindGloblException();
 // eslint-disable-next-line no-multi-assign
 GameGlobal.onCrash = GameGlobal.unityNamespace.onCrash = function () {
     GameGlobal.manager.showAbort();
-    const windowInfo = wx.getWindowInfo();
+    const windowInfo = wx.getWindowInfo ? wx.getWindowInfo() : wx.getSystemInfoSync();
     wx.createFeedbackButton({
         type: 'text',
         text: '提交反馈',