+++ Runtime/wechat-default/unity-sdk/video/index.js
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/prefer-for-of */
 /* eslint-disable @typescript-eslint/naming-convention */
-import { isH5Renderer, isSupportVideoPlayer, isPc, isDevtools } from '../../check-version';
+import { isH5Renderer, isSupportVideoPlayer, isDebug, isPc, isDevtools } from '../../check-version';
 let FrameworkData = null;
 const isWebVideo = isH5Renderer || isPc || isDevtools;
-const isDebug = false;
 const needCache = true;
 const cacheVideoDecoder = [];
 const videoInstances = {};